[up:objective-functions-losses] refactor: better file structure for nujo objectives
