+++ nujo/objective/__init__.py
@@ -1,0 +1,2 @@
+from nujo.objective.qualitative import *
+from nujo.objective.quantitative import *
